133 - rename v2 sequencer to base sequencer
