+++ tests/tel2puml/otel_to_pv/test_sequence_otel.py
@@ -7,7 +7,7 @@ import yaml
 import pytest
 from pytest import MonkeyPatch
 
-from tel2puml.otel_to_pv.sequence_otel_v2 import (
+from tel2puml.otel_to_pv.sequence_otel import (
     order_groups_by_start_timestamp,
     sequence_groups_of_otel_events_asynchronously,
     group_events_using_async_information,
